pyproject.toml
[project]
name = "akta"
version = "0.1.0"
description = "Akta is a library for managing DIDs, VCs, and VDRs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "base58>=2.1.1",
    "click>=8.2.1",
    "pydantic-settings>=2.9.1",
    "pynacl>=1.5.0",
    "pyld>=2.0.4",
    "httpx>=0.28.1",
    "pyakta",
    "uvicorn>=0.34.3",
    "fastapi>=0.115.12",
    "sqlalchemy>=2.0.41",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["akta/*"]

[tool.uv.sources]
pyakta = { path = "../pyakta", editable = true }

[project.scripts]
akta = "akta.cli:cli"
